requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
motor==3.5.1
pydantic==2.9.2
pydantic-settings==2.6.1
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
telethon==1.36.0
qrcode==7.4.2
Pillow==10.4.0
